// client/src/api/models/DocumentRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Knowledge Manager API
 * The Knowledge Manager API provides endpoints for managing knowledge units. A knowledge unit aggregates multi-dimensional information needed for specific tasks, enabling cross-domain information reuse and contextual knowledge assembly.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DocumentRequest
 */
export interface DocumentRequest {
    /**
     * 
     * @type {string}
     * @memberof DocumentRequest
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof DocumentRequest
     */
    content?: string;
}

/**
 * Check if a given object implements the DocumentRequest interface.
 */
export function instanceOfDocumentRequest(value: object): value is DocumentRequest {
    return true;
}

export function DocumentRequestFromJSON(json: any): DocumentRequest {
    return DocumentRequestFromJSONTyped(json, false);
}

export function DocumentRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): DocumentRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
        'content': json['content'] == null ? undefined : json['content'],
    };
}

export function DocumentRequestToJSON(json: any): DocumentRequest {
    return DocumentRequestToJSONTyped(json, false);
}

export function DocumentRequestToJSONTyped(value?: DocumentRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'content': value['content'],
    };
}